Resolve (glob) configuration paths.
